Refactor grammar.rb that separate Code class from lrama/grammar.rb
